max和min修改成BigDecimal

+++ aggregate-function-impl/src/test/java/com/yanggu/metric_calculate/core/aggregate_function/numeric/MinAggregateFunctionTest.java
@@ -3,6 +3,8 @@ package com.yanggu.metric_calculate.core.aggregate_function.numeric;
 import com.yanggu.metric_calculate.core.aggregate_function.AggregateFunctionTestBase;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class MinAggregateFunctionTest {
@@ -25,57 +27,57 @@ class MinAggregateFunctionTest {
     @Test
     void testCreateAccumulator() {
         MinAggregateFunction<Integer> minAggregateFunction = new MinAggregateFunction<>();
-        Double accumulator = minAggregateFunction.createAccumulator();
-        assertEquals(Double.MAX_VALUE, accumulator, 0.0D);
+        BigDecimal accumulator = minAggregateFunction.createAccumulator();
+        assertEquals(Double.MAX_VALUE, accumulator.doubleValue(), 0.0D);
     }
 
     @Test
     void testAddPositive() {
         MinAggregateFunction<Integer> minAggregateFunction = new MinAggregateFunction<>();
-        Double accumulator = minAggregateFunction.createAccumulator();
-        Double result = minAggregateFunction.add(1, accumulator);
-        assertEquals(1.0D, result, 0.0D);
+        BigDecimal accumulator = minAggregateFunction.createAccumulator();
+        BigDecimal result = minAggregateFunction.add(1, accumulator);
+        assertEquals(1.0D, result.doubleValue(),  0.0D);
     }
 
     @Test
     void testAddNegative() {
         MinAggregateFunction<Integer> minAggregateFunction = new MinAggregateFunction<>();
-        Double accumulator = minAggregateFunction.createAccumulator();
-        Double result = minAggregateFunction.add(-1, accumulator);
-        assertEquals(-1.0D, result, 0.0D);
+        BigDecimal accumulator = minAggregateFunction.createAccumulator();
+        BigDecimal result = minAggregateFunction.add(-1, accumulator);
+        assertEquals(-1.0D, result.doubleValue(),  0.0D);
     }
 
     @Test
     void testGetResultPositive() {
         MinAggregateFunction<Integer> minAggregateFunction = new MinAggregateFunction<>();
-        Double accumulator = minAggregateFunction.createAccumulator();
-        Double result = minAggregateFunction.getResult(accumulator);
-        assertEquals(Double.MAX_VALUE, result, 0.0D);
+        BigDecimal accumulator = minAggregateFunction.createAccumulator();
+        BigDecimal result = minAggregateFunction.getResult(accumulator);
+        assertEquals(Double.MAX_VALUE, result.doubleValue(),  0.0D);
     }
 
     @Test
     void testGetResultNegative() {
         MinAggregateFunction<Integer> minAggregateFunction = new MinAggregateFunction<>();
-        Double accumulator = minAggregateFunction.add(-1, minAggregateFunction.createAccumulator());
-        Double result = minAggregateFunction.getResult(accumulator);
-        assertEquals(-1.0D, result, 0.0D);
+        BigDecimal accumulator = minAggregateFunction.add(-1, minAggregateFunction.createAccumulator());
+        BigDecimal result = minAggregateFunction.getResult(accumulator);
+        assertEquals(-1.0D, result.doubleValue(),  0.0D);
     }
 
     @Test
     void testMergePositive() {
         MinAggregateFunction<Integer> minAggregateFunction = new MinAggregateFunction<>();
-        Double thisAccumulator = minAggregateFunction.add(1, minAggregateFunction.createAccumulator());
-        Double thatAccumulator = minAggregateFunction.add(2, minAggregateFunction.createAccumulator());
-        Double result = minAggregateFunction.merge(thisAccumulator, thatAccumulator);
-        assertEquals(1.0D, result, 0.0D);
+        BigDecimal thisAccumulator = minAggregateFunction.add(1, minAggregateFunction.createAccumulator());
+        BigDecimal thatAccumulator = minAggregateFunction.add(2, minAggregateFunction.createAccumulator());
+        BigDecimal result = minAggregateFunction.merge(thisAccumulator, thatAccumulator);
+        assertEquals(1.0D, result.doubleValue(),  0.0D);
     }
 
     @Test
     void testMergeNegative() {
         MinAggregateFunction<Integer> minAggregateFunction = new MinAggregateFunction<>();
-        Double thisAccumulator = minAggregateFunction.add(-1, minAggregateFunction.createAccumulator());
-        Double thatAccumulator = minAggregateFunction.add(-2, minAggregateFunction.createAccumulator());
-        Double result = minAggregateFunction.merge(thisAccumulator, thatAccumulator);
-        assertEquals(-2.0D, result, 0.0D);
+        BigDecimal thisAccumulator = minAggregateFunction.add(-1, minAggregateFunction.createAccumulator());
+        BigDecimal thatAccumulator = minAggregateFunction.add(-2, minAggregateFunction.createAccumulator());
+        BigDecimal result = minAggregateFunction.merge(thisAccumulator, thatAccumulator);
+        assertEquals(-2.0D, result.doubleValue(),  0.0D);
     }
 }
